modules/watcher/requirements.txt
requests
feedparser
python-dotenv
yt-dlp
pinecone-client